app/src/main/java/mega/privacy/android/app/lollipop/AudioVideoPlayerLollipop.java
app/src/main/java/mega/privacy/android/app/lollipop/FullScreenImageViewerLollipop.java
app/src/main/java/mega/privacy/android/app/lollipop/ManagerActivityLollipop.java
app/src/main/java/mega/privacy/android/app/lollipop/TestPasswordActivity.java
app/src/main/java/mega/privacy/android/app/lollipop/TwoFactorAuthenticationActivity.java
app/src/main/java/mega/privacy/android/app/lollipop/adapters/MegaOfflineLollipopAdapter.java
app/src/main/java/mega/privacy/android/app/lollipop/controllers/AccountController.java
app/src/main/java/mega/privacy/android/app/lollipop/controllers/NodeController.java
app/src/main/java/mega/privacy/android/app/lollipop/managerSections/MyAccountFragmentLollipop.java
app/src/main/java/mega/privacy/android/app/lollipop/managerSections/OfflineFragmentLollipop.java
app/src/main/java/mega/privacy/android/app/modalbottomsheet/OfflineOptionsBottomSheetDialogFragment.java

+++ app/src/main/java/mega/privacy/android/app/lollipop/ManagerActivityLollipop.java
@@ -67,7 +67,6 @@ import android.text.InputType;
 import android.text.Spanned;
 import android.text.TextWatcher;
 import android.util.DisplayMetrics;
-import android.util.Log;
 import android.util.TypedValue;
 import android.view.ActionMode;
 import android.view.Display;
@@ -196,7 +195,6 @@ import mega.privacy.android.app.modalbottomsheet.SentRequestBottomSheetDialogFra
 import mega.privacy.android.app.modalbottomsheet.TransfersBottomSheetDialogFragment;
 import mega.privacy.android.app.modalbottomsheet.UploadBottomSheetDialogFragment;
 import mega.privacy.android.app.modalbottomsheet.chatmodalbottomsheet.ChatBottomSheetDialogFragment;
-import mega.privacy.android.app.utils.CacheFolderManager;
 import mega.privacy.android.app.utils.ChatUtil;
 import mega.privacy.android.app.utils.Constants;
 import mega.privacy.android.app.utils.JobUtil;
@@ -243,15 +241,16 @@ import nz.mega.sdk.MegaUserAlert;
 import nz.mega.sdk.MegaUtilsAndroid;
 
 import static mega.privacy.android.app.lollipop.FileInfoActivityLollipop.NODE_HANDLE;
-import static mega.privacy.android.app.utils.CacheFolderManager.buildAvatarFile;
-import static mega.privacy.android.app.utils.CacheFolderManager.buildQrFile;
-import static mega.privacy.android.app.utils.CacheFolderManager.isFileAvailable;
 import static mega.privacy.android.app.utils.Constants.*;
+import static mega.privacy.android.app.utils.CacheFolderManager.*;
+import static mega.privacy.android.app.utils.FileUtils.*;
 import static mega.privacy.android.app.utils.JobUtil.stopRunningCameraUploadService;
 import static mega.privacy.android.app.utils.Util.showSnackBar;
 
 public class ManagerActivityLollipop extends PinActivityLollipop implements MegaRequestListenerInterface, MegaChatListenerInterface, MegaChatCallListenerInterface,MegaChatRequestListenerInterface, OnNavigationItemSelectedListener, MegaGlobalListenerInterface, MegaTransferListenerInterface, OnClickListener,
 			NodeOptionsBottomSheetDialogFragment.CustomHeight, ContactsBottomSheetDialogFragment.CustomHeight, View.OnFocusChangeListener, View.OnLongClickListener, BottomNavigationView.OnNavigationItemSelectedListener {
+
+	private final String MK_LAYOUT_VISIBLE = "MK_LAYOUT_VISIBLE";
 
 	final int CLOUD_DRIVE_BNV = 0;
 	final int CAMERA_UPLOADS_BNV = 1;
@@ -1904,7 +1903,7 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
 		cC = new ContactController(this);
 		aC = new AccountController(this);
 
-        CacheFolderManager.createCacheFolders(this);
+        createCacheFolders(this);
 
         dbH = DatabaseHandler.getDbHandler(getApplicationContext());
 
@@ -2350,13 +2349,10 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
 		///Check the MK file
 		int versionApp = Util.getVersion(this);
 		log("-------------------Version app: "+versionApp);
-		final String pathOldMK = Environment.getExternalStorageDirectory().getAbsolutePath()+Util.oldMKFile;
-		final File fMKOld = new File(pathOldMK);
-		if (fMKOld != null) {
-			if (fMKOld.exists()) {
-				log("Old MK file need to be renamed!");
-				aC.renameMK();
-			}
+		final File fMKOld = buildExternalStorageFile(OLD_MK_FILE);
+		if (isFileAvailable(fMKOld)) {
+			log("Old MK file need to be renamed!");
+			aC.renameMK();
 		}
 
 		rootNode = megaApi.getRootNode();
@@ -11120,12 +11116,7 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
 
 	public void takePicture(){
 		log("takePicture");
-		String path = Environment.getExternalStorageDirectory().getAbsolutePath() +"/"+ Util.temporalPicDIR;
-        File newFolder = new File(path);
-        newFolder.mkdirs();
-
-        String file = path + "/picture.jpg";
-        File newFile = new File(file);
+        File newFile = buildTempFile(this, "picture.jpg");
         try {
         	newFile.createNewFile();
         } catch (IOException e) {}
@@ -11174,13 +11165,7 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
 	}
 
 	public void takeProfilePicture(){
-		log("takeProfilePicture");
-		String path = Environment.getExternalStorageDirectory().getAbsolutePath() +"/"+ Util.profilePicDIR;
-		File newFolder = new File(path);
-		newFolder.mkdirs();
-
-		String file = path + "/picture.jpg";
-		File newFile = new File(file);
+		File newFile = buildTempFile(this, "picture.jpg");
 		try {
 			newFile.createNewFile();
 		} catch (IOException e) {}
@@ -13542,12 +13527,12 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
 				Intent service = new Intent(this, DownloadService.class);
 				service.putExtra(DownloadService.EXTRA_HASH, handleToDownload);
 				service.putExtra(DownloadService.EXTRA_CONTENT_URI, treeUri.toString());
-				String path = Environment.getExternalStorageDirectory().getAbsolutePath() + "/" + Util.advancesDevicesDIR + "/";
-				File tempDownDirectory = new File(path);
-				if(!tempDownDirectory.exists()){
-					tempDownDirectory.mkdirs();
-				}
-				service.putExtra(DownloadService.EXTRA_PATH, path);
+				File tempFolder = getCacheFolder(this, TEMPORAL_FOLDER);
+				if (!isFileAvailable(tempFolder)) {
+				    showSnackbar(Constants.SNACKBAR_TYPE, getString(R.string.general_error), -1);
+				    return;
+                }
+				service.putExtra(DownloadService.EXTRA_PATH, tempFolder.getAbsolutePath());
 				startService(service);
 			}
 		}
@@ -13938,17 +13923,18 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
 		else if (requestCode == TAKE_PHOTO_CODE){
 			log("TAKE_PHOTO_CODE");
 			if(resultCode == Activity.RESULT_OK){
-				String filePath = Environment.getExternalStorageDirectory().getAbsolutePath() +"/"+ Util.temporalPicDIR + "/picture.jpg";
-				File imgFile = new File(filePath);
+				File imgFile = getCacheFile(this, TEMPORAL_FOLDER, "picture.jpg");
+				if (!isFileAvailable(imgFile)) {
+					showSnackbar(Constants.SNACKBAR_TYPE, getString(R.string.general_error), -1);
+					return;
+				}
 
 				String name = Util.getPhotoSyncName(imgFile.lastModified(), imgFile.getAbsolutePath());
 				log("Taken picture Name: "+name);
-				String newPath = Environment.getExternalStorageDirectory().getAbsolutePath() +"/"+ Util.temporalPicDIR + "/"+name;
-				log("----NEW Name: "+newPath);
-				File newFile = new File(newPath);
+				File newFile = buildTempFile(this, name);
 				imgFile.renameTo(newFile);
 
-				uploadTakePicture(newPath);
+				uploadTakePicture(newFile.getAbsolutePath());
 			}
 			else{
 				log("TAKE_PHOTO_CODE--->ERROR!");
@@ -13960,8 +13946,11 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
 			if(resultCode == Activity.RESULT_OK){
 
 				String myEmail =  megaApi.getMyUser().getEmail();
-				String filePath = Environment.getExternalStorageDirectory().getAbsolutePath() +"/"+ Util.profilePicDIR + "/picture.jpg";
-				File imgFile = new File(filePath);
+				File imgFile = getCacheFile(this, TEMPORAL_FOLDER, "picture.jpg");
+				if (!isFileAvailable(imgFile)) {
+					showSnackbar(Constants.SNACKBAR_TYPE, getString(R.string.general_error), -1);
+					return;
+				}
 
                 File qrFile = buildQrFile(this,myEmail + "QRcode.jpg");
                 File newFile = buildAvatarFile(this,myEmail + "Temp.jpg");
@@ -14106,7 +14095,8 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
 		else if (requestCode == REQUEST_DOWNLOAD_FOLDER && resultCode == RESULT_OK){
 			String parentPath = intent.getStringExtra(FileStorageActivityLollipop.EXTRA_PATH);
 			if (parentPath != null){
-				String path = parentPath + File.separator + Util.rKFile;
+				String path = parentPath + File.separator + RK_FILE;
+
 				log("REQUEST_DOWNLOAD_FOLDER:path to download: "+path);
 				AccountController ac = new AccountController(this);
 				ac.exportMK(path);
@@ -15064,13 +15054,6 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
                         } else {
                             log("ERROR! Destination PATH is NULL");
                         }
-
-
-//						String newPath = Environment.getExternalStorageDirectory().getAbsolutePath() +"/"+ Util.profilePicDIR + "/"+name;
-//						log("----NEW Name: "+newPath);
-//						File newFile = new File(newPath);
-//						MegaUtilsAndroid.createAvatar(imgFile, newFile);
-
 					}
 					else{
 						log("The chosen avatar path is NULL");
@@ -15186,7 +15169,7 @@ public class ManagerActivityLollipop extends PinActivityLollipop implements Mega
 
 	private void createFile(String name, String data, MegaNode parentNode){
 
-		File file = Util.createTemporalTextFile(name, data);
+		File file = createTemporalTextFile(this, name, data);
 		if(file!=null){
 			showSnackbar(SNACKBAR_TYPE, getString(R.string.upload_began), -1);
